requirements.txt
streamlit
faiss-cpu
torch
numpy
pandas
transformers
watchdog